Split dataset loading to different files
